ARQ-1653 Add Thread propagation of Arquillian Contexts
The ExecutorService used in DroneInstanceCreator attempts to create
new Driver instances on a new Thread. The new Thread is missing the
Arquillian Contexts as they are bound to ThreadLocals. We need to
make sure the active Contexts used during the call to
DroneInstanceCreator is reactivated on the new Thread when creating
the Driver instance.

This addresses the Driver instance creation problem introduced by
Core 1.1.3.Final after reverting the Thread inheritance introduced
in 1.0.4.Final.