ARQ-2037 Added a support of additional browser capabilities for Chrome
With this, it should be possible to set all possible chrome options into
a chromeDriver through arquillian.xml file. All the options you can find
here: https://sites.google.com/a/chromium.org/chromedriver/capabilities
All the options are set into ChromeOptions class:
https://seleniumhq.github.io/selenium/docs/api/java/org/openqa/selenium/chrome/ChromeOptions.html
which means that the parameter names tightly depend on names of the set/add
methods.
It is expected that the name of a parameter consist of:
"chrome" + (name of set/add method of ChromeOption class without first three chars)
whole string in camelcase.
eg for option args there should be:
<property name="chromeArguments">--your-cool --arguments</property>
If the value can be an array or list of strings/files, then specify all of
them in one string separated by space. (This is also applied for
extensions as well as for encoded extensions.)
In the case of experimental options it is a little bit different. As there
is expected a set of dictionaries, the most suitable way is using JSON
format - eg:
<property name="chromeExperimentalOption">{"perfLoggingPrefs": {
"traceCategories":
",blink.console,disabled-by-default-devtools.timeline,benchmark"
 }, "prefs": {"download.default_directory":
"/usr/local/path/to/download/directory"} }</property>
(can be in multiline format )

If you are still struglling with passing required options through
arquillian.xml, you can use a parameter chromePrintOptions with a value
"true":
<property name="chromePrintOptions">true</property>
This ensures that Drone prints whole content of ChromeOptions in a JSON
format on a standard output.